+++ src/ionic-tree-provider.ts
@@ -33,6 +33,7 @@ interface IonicState {
   refreshDebugDevices: boolean; // Should we refresh the list of debuggable devices
   remoteLogging: boolean; // Whether remote logging is enabled
   hasNodeModules: boolean; // Whether node modules are installed
+  nodeModulesFolder: string; // The folder where node_modules is located
   hasPackageJson: boolean; // Whether folder has package.json
   hasNodeModulesNotified: boolean; // Whether we've notified the user of no node_modules
   configuration: string; // Build configuration
@@ -41,6 +42,7 @@ interface IonicState {
   rootFolder: string; // The folder to inspect
   flavors: string[]; // Android Flavors
   debugged: boolean; // Have we ever started debugging
+  servePort: number; // The port used when the dev server is running
   runIOS: Tip;
   runAndroid: Tip;
   runWeb: Tip;
@@ -59,6 +61,7 @@ export const ionicState: IonicState = {
   outputIsFocused: false,
   channelFocus: true,
   hasNodeModules: undefined,
+  nodeModulesFolder: undefined,
   hasPackageJson: undefined,
   hasNodeModulesNotified: undefined,
   syncDone: [],
@@ -73,6 +76,7 @@ export const ionicState: IonicState = {
   lastRun: undefined,
   projectRef: undefined,
   configuration: undefined,
+  servePort: 8100,
   project: undefined,
   debugged: false,
 };